services/frappe_service/requirements.txt
frappe-bench>=0.0.1
builder>=0.0.1
gunicorn>=21.2.0
werkzeug>=2.3.0

